+++ src/leaflet-layer.js
@@ -15,6 +15,7 @@ export default class LeafletLayer extends L.Layer {
   _deck = undefined;
 
   updateBound = this.update.bind(this);
+  animateZoomBound = this.animateZoom.bind(this);
 
   /**
    * @param {DeckProps} props
@@ -34,11 +35,16 @@ export default class LeafletLayer extends L.Layer {
     this._container = L.DomUtil.create('div');
     this._container.className = 'leaflet-layer';
     this._container.style.pointerEvents = 'none';
+    if (this._zoomAnimated) {
+      L.DomUtil.addClass(this._container, 'leaflet-zoom-animated');
+    }
     this.getPane().appendChild(this._container);
+
     this._deck = createDeckInstance(this._map, this._container, this._deck, this.props);
 
     this._map.on('resize', this.updateBound);
     this._map.on('move', this.updateBound);
+    this._map.on('zoomanim', this.animateZoomBound);
     this._map.on('zoom', this.updateBound);
 
     this.update();
@@ -53,6 +59,7 @@ export default class LeafletLayer extends L.Layer {
   onRemove(_map) {
     this._map.off('resize', this.updateBound);
     this._map.off('move', this.updateBound);
+    this._map.off('zoomanim', this.animateZoomBound);
     this._map.off('zoom', this.updateBound);
 
     this._map = undefined;
@@ -84,9 +91,40 @@ export default class LeafletLayer extends L.Layer {
     this._container.style.height = `${this._map.getSize().y}px`;
 
     // invert map position
-    const offset = L.DomUtil.getPosition(this._map.getPanes().mapPane).multiplyBy(-1);
+    const offset = this._map._getMapPanePos().multiplyBy(-1);
     L.DomUtil.setPosition(this._container, offset);
 
     updateDeckView(this._deck, this._map);
+  }
+
+  /**
+   * @param {L.ZoomAnimEvent} event
+   * @returns {void}
+   */
+  animateZoom(event) {
+    this.updateTransform(event.center, event.zoom);
+  }
+
+  /**
+   * see https://stackoverflow.com/a/67107000/1823988
+   * see L.Renderer._updateTransform https://github.com/Leaflet/Leaflet/blob/master/src/layer/vector/Renderer.js#L90-L105
+   * @param {L.LatLng} center
+   * @param {number} zoom
+   */
+  updateTransform(center, zoom) {
+    var scale = this._map.getZoomScale(zoom, this._map.getZoom()),
+        position = L.DomUtil.getPosition(this._container),
+        viewHalf = this._map.getSize().multiplyBy(0.5),
+        currentCenterPoint = this._map.project(this._map.getCenter(), zoom),
+        destCenterPoint = this._map.project(center, zoom),
+        centerOffset = destCenterPoint.subtract(currentCenterPoint),
+
+        topLeftOffset = viewHalf.multiplyBy(-scale).add(position).add(viewHalf).subtract(centerOffset);
+
+    if (L.Browser.any3d) {
+      L.DomUtil.setTransform(this._container, topLeftOffset, scale);
+    } else {
+      L.DomUtil.setPosition(this._container, topLeftOffset);
+    }
   }
 }